Rename pipeline stats.php
