+++ includes/pipeline_page/pipeline_stats.php
